+++ evaluators.py
@@ -5,20 +5,22 @@ from multiprocessing import Pool, cpu_count
 from gym import Env
 import numpy as np
 
-from utils import PolicyInterface, ObsNormalizer
+from utils import Policy, ObsNormalizer
+
 
 class Evaluator(ABC):
     @abstractmethod
-    def __call__(self, policies: List[PolicyInterface]) -> List[Tuple[float, PolicyInterface]]:
+    def __call__(self, policies: List[Policy]) -> List[Tuple[float, Policy]]:
         pass
+
 
 class EnvEvaluator(Evaluator):
     def __init__(self, env_factory: function):
         self.env: Env = env_factory()
         self.normalizer = ObsNormalizer(env_factory)
 
-    def __call__(self, policies: List[PolicyInterface], times=1) -> List[Tuple[float, PolicyInterface]]:
-        results: List[Tuple[float, PolicyInterface]] = []
+    def __call__(self, policies: List[Policy], times=1) -> List[Tuple[float, Policy]]:
+        results: List[Tuple[float, Policy]] = []
         for policy in policies:
             rewards = []
             for _ in range(times):
@@ -34,18 +36,19 @@ class EnvEvaluator(Evaluator):
             results.append((int(np.mean(rewards)), policy))
         return results
 
+
 class ParallelEnvEvaluator(Evaluator):
     def __init__(self, env_factory: function):
         self.env_factory = env_factory
         self.normalizer = ObsNormalizer(env_factory)
 
-    def __call__(self, policies: List[PolicyInterface], times=1) -> List[Tuple[float, PolicyInterface]]:
+    def __call__(self, policies: List[Policy], times=1) -> List[Tuple[float, Policy]]:
         with Pool(cpu_count()) as p:
             # List[Tuple[float, PolicyInterface]]
             results = p.starmap(self._eval_policy, policies)
         return results
 
-    def _eval_policy(self, policy: PolicyInterface, times=1):
+    def _eval_policy(self, policy: Policy, times=1):
         env: Env = self.env_factory()
         rewards = []
         for _ in range(times):
@@ -59,4 +62,3 @@ class ParallelEnvEvaluator(Evaluator):
                 total_reward += reward
             rewards.append(total_reward)
         return int(np.mean(rewards)), policy
-
