+++ halfcheetah_new.py
@@ -3,4 +3,4 @@ import numpy
 import torch
 
 from evolution_strategies import BasicStrategy
-from dne_optimizer import DeepEvolution
+from deep_neuroevolution import GAOptimizer
